refactor: move strategies to new file
